fix: smaller task fixes
