+++ shared/types/mention.ts
@@ -4,22 +4,19 @@ export const REQUEST_RESPONSE = "request-response";
 
 export const REQUEST_ACTION = "request-action";
 
+export const REQUEST_TYPES = [REQUEST_READ, REQUEST_RESPONSE, REQUEST_ACTION] as const;
+
+export type RequestType = typeof REQUEST_TYPES[number];
+
 export const MENTION_OBSERVER = "observer";
 
-export type RequestType = typeof REQUEST_READ | typeof REQUEST_RESPONSE | typeof REQUEST_ACTION;
+export const MENTION_TYPES = [...REQUEST_TYPES, MENTION_OBSERVER] as const;
 
-export type MentionType = RequestType | typeof MENTION_OBSERVER;
+export type MentionType = typeof MENTION_TYPES[number];
 
 export const MENTION_TYPE_LABELS: Record<MentionType, string> = {
-  [REQUEST_RESPONSE]: "Feedback",
+  [REQUEST_RESPONSE]: "Response",
   [REQUEST_READ]: "Read Confirmation",
   [REQUEST_ACTION]: "Action",
   [MENTION_OBSERVER]: "Observer",
-};
-
-export const NEW_MENTION_TYPE_LABELS: Record<MentionType, string> = {
-  [REQUEST_RESPONSE]: "Request response",
-  [REQUEST_READ]: "Request read confirmation",
-  [REQUEST_ACTION]: "Request action",
-  [MENTION_OBSERVER]: "Add as observer",
 };
